feat: Refactor Dashboard component; move ImhotepRecipe and IngredientsList to separate files for better organization
